# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Cloud-Based Restaurant Digital Menu Management System API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export interface LoginInput {
  username: string;
  password: string;
}

export interface AdminUser {
  id: number;
  username: string;
}

export interface LoginResponse {
  success: boolean;
  admin: AdminUser;
}

export type RestaurantSubscriptionStatus =
  (typeof RestaurantSubscriptionStatus)[keyof typeof RestaurantSubscriptionStatus];

export const RestaurantSubscriptionStatus = {
  active: "active",
  inactive: "inactive",
  suspended: "suspended",
} as const;

export interface Restaurant {
  id: number;
  /** Unique customer ID like REST001 */
  customerId: string;
  name: string;
  /** @nullable */
  contactEmail?: string | null;
  /** @nullable */
  contactPhone?: string | null;
  /** @nullable */
  address?: string | null;
  subscriptionStatus: RestaurantSubscriptionStatus;
  /** @nullable */
  activeMenuId?: number | null;
  /** @nullable */
  activeMenuUrl?: string | null;
  createdAt: string;
}

export interface RestaurantInput {
  name: string;
  contactEmail?: string;
  contactPhone?: string;
  address?: string;
}

export interface RestaurantUpdate {
  name?: string;
  contactEmail?: string;
  contactPhone?: string;
  address?: string;
}

export type SubscriptionUpdateStatus =
  (typeof SubscriptionUpdateStatus)[keyof typeof SubscriptionUpdateStatus];

export const SubscriptionUpdateStatus = {
  active: "active",
  inactive: "inactive",
  suspended: "suspended",
} as const;

export interface SubscriptionUpdate {
  status: SubscriptionUpdateStatus;
}

export interface RestaurantStats {
  total: number;
  active: number;
  inactive: number;
  suspended: number;
  recentUploads: number;
}

export interface Menu {
  id: number;
  restaurantId: number;
  /** @nullable */
  restaurantName?: string | null;
  /** @nullable */
  restaurantCustomerId?: string | null;
  imageUrl: string;
  /** @nullable */
  notes?: string | null;
  isActive: boolean;
  uploadedAt: string;
}

export interface MenuUploadInput {
  restaurantId: number;
  notes?: string;
}

export interface DisplayMenu {
  restaurantId: number;
  customerId: string;
  restaurantName: string;
  imageUrl: string;
  updatedAt: string;
}

export type ListMenusParams = {
  restaurantId?: number;
};