# lib/api-zod/src/generated/types/displayMenu.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Cloud-Based Restaurant Digital Menu Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface DisplayMenu {
  restaurantId: number;
  customerId: string;
  restaurantName: string;
  imageUrl: string;
  updatedAt: string;
}