# lib/api-zod/src/generated/types/restaurant.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Cloud-Based Restaurant Digital Menu Management System API
 * OpenAPI spec version: 0.1.0
 */
import type { RestaurantSubscriptionStatus } from "./restaurantSubscriptionStatus";

export interface Restaurant {
  id: number;
  /** Unique customer ID like REST001 */
  customerId: string;
  name: string;
  /** @nullable */
  contactEmail?: string | null;
  /** @nullable */
  contactPhone?: string | null;
  /** @nullable */
  address?: string | null;
  subscriptionStatus: RestaurantSubscriptionStatus;
  /** @nullable */
  activeMenuId?: number | null;
  /** @nullable */
  activeMenuUrl?: string | null;
  createdAt: string;
}