# lib/api-zod/src/generated/types/subscriptionUpdate.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Cloud-Based Restaurant Digital Menu Management System API
 * OpenAPI spec version: 0.1.0
 */
import type { SubscriptionUpdateStatus } from "./subscriptionUpdateStatus";

export interface SubscriptionUpdate {
  status: SubscriptionUpdateStatus;
}